frontend/site-license-info: refactoring large tsx file and removing an extra gap in license column under the description
