+++ src/packages/frontend/site-licenses/util.ts
@@ -8,7 +8,7 @@ import { reuseInFlight } from "async-await-utils/hof";
 import { SiteLicensePublicInfo } from "./types";
 import { query } from "../frame-editors/generic/client";
 import { SCHEMA } from "@cocalc/util/db-schema";
-import { copy } from "@cocalc/util/misc";
+import { copy, trunc_left } from "@cocalc/util/misc";
 
 // To avoid overfetching, we cache results for *a few seconds*.
 const site_license_public_info_cache = new LRU({ maxAge: 1000 * 15 });
@@ -39,3 +39,7 @@ export const site_license_public_info = reuseInFlight(async function (
   site_license_public_info_cache.set(license_id, info);
   return info;
 });
+
+export function trunc_license_id(license_id: string) {
+  return trunc_left(license_id, 13);
+}
